Preliminary BC: 2e5

+++ BC_CAF2021_Masatepe_01-14.R
@@ -2,8 +2,7 @@
 ## MvO, 2021-05-20
 
 ## 1. INITIALISE MCMC ##
-   nChain <- as.integer(2e2)
-   # nChain <- as.integer(1e5)
+   nChain <- as.integer(2e5)
    source('BC/BC_CAF2021_MCMC_init_Masatepe_01-14.R')
 
 ## 2. RUNNING THE MCMC ##
